requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pytest>=6.2.0
tkcalendar>=1.6.0
bcrypt>=3.2.0
pillow>=8.2.0
